requirements.txt
Flask==3.0.1
python-dotenv==1.0.1
pytest==8.3.3
numpy==2.1.3
